refactor(server): eliminate circular deps, switch to bun build, slim deps
- Extract gateway singleton and voice connection into lib/gatewayState.ts,
  breaking the startDiscord -> manager -> GuildPlayer -> startDiscord cycle
- Convert all 7 dynamic imports to static imports
- Switch server build from tsc to bun build (19ms vs 470ms)
- Remove typescript, @types/node, bun-types from dependency tree
- Remove stale lodash/serialize-javascript overrides
- Move drizzle-orm to devDependencies (fully bundled)
- Replace __dirname/__filename with import.meta.dir/.filename
- Delete dead root tsconfig.json
- Drop tsc --noEmit from web build
- Simplify Dockerfile CMD
