+++ packages/server/package.json
@@ -1,23 +1,19 @@
 {
-  "name": "@alfira-bot/server",
+  "name": "@alfira/server",
   "version": "0.1.1",
   "private": true,
-  "main": "./dist/index.js",
+  "main": "./src/index.ts",
   "exports": {
-    ".": "./dist/index.js",
-    "./shared": "./dist/shared/index.js",
-    "./shared/*": "./dist/shared/*"
+    ".": "./src/index.ts",
+    "./shared": "./src/shared/index.ts",
+    "./shared/*": "./src/shared/*"
   },
   "scripts": {
     "dev": "bun run src/index.ts",
-    "build": "tsc --skipLibCheck"
+    "build": "bun build ./src/index.ts --outdir=./dist --target bun --format esm"
   },
-  "dependencies": {
-    "drizzle-orm": "0.45.2"
-  },
+  "dependencies": {},
   "devDependencies": {
-    "@types/node": "25.9.4",
-    "bun-types": "1.3.14",
-    "typescript": "7.0.2"
+    "drizzle-orm": "0.45.2"
   }
 }
