+++ packages/web/package.json
@@ -1,15 +1,15 @@
 {
-  "name": "@alfira-bot/web",
+  "name": "@alfira/web",
   "version": "0.1.1",
   "private": true,
   "type": "module",
   "scripts": {
     "dev": "bunx @tailwindcss/cli -i ./src/index.css -o ./public/index.css --watch & bun run src/server.ts",
     "build:css": "bunx @tailwindcss/cli -i ./src/index.css -o ./dist/index.css --minify",
-    "build": "tsc --noEmit && bun run build:css && bun build ./src/main.tsx --outdir=./dist --minify --sourcemap=external && cp ./index.html ./dist/ && sed -i 's|/src/main.tsx|/main.js|g' ./dist/index.html && sed -i 's|/src/index.css|/index.css|g' ./dist/index.html && cp -r ./public/* ./dist/"
+    "build": "bun run build:css && bun build ./src/main.tsx --outdir=./dist --minify --sourcemap=external && cp ./index.html ./dist/ && sed -i 's|/src/main.tsx|/main.js|g' ./dist/index.html && sed -i 's|/src/index.css|/index.css|g' ./dist/index.html && cp -r ./public/* ./dist/"
   },
   "dependencies": {
-    "@alfira-bot/server": "workspace:*",
+    "@alfira/server": "workspace:*",
     "@phosphor-icons/react": "2.1.10",
     "@tanstack/react-virtual": "3.14.6",
     "react": "19.2.7",
@@ -20,8 +20,6 @@
     "@tailwindcss/cli": "4.3.2",
     "@types/react": "19.2.17",
     "@types/react-dom": "19.2.3",
-    "bun-types": "1.3.14",
-    "tailwindcss": "4.3.2",
-    "typescript": "7.0.2"
+    "tailwindcss": "4.3.2"
   }
 }
